requirements.txt
asgiref==3.2.10
autopep8==1.5.4
click==7.1.2
dj-database-url==0.5.0
Django==3.1.1
django-cors-headers==3.5.0
django-heroku==0.3.1
djangorestframework==3.11.1
future==0.18.2
gunicorn==20.0.4
joblib==0.16.0
nltk==3.5
numpy==1.19.2
psycopg2==2.8.6
pycodestyle==2.6.0
pytz==2020.1
regex==2020.7.14
sqlparse==0.3.1
toml==0.10.1
torch==1.6.0
tqdm==4.48.2
whitenoise==5.2.0
